% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucor.R
\docType{package}
\name{accucor}
\alias{accucor-package}
\alias{accucor}
\title{accucor: A package for natural abundance correction of mass spectrometer data}
\description{
AccuCor is an isotope natural abundance correction algorithm that
is needed especially for high resolution mass spectrometers. AccuCor
supports correction for 13C, 2H and 15N.
}
\section{AccuCor functions}{

\code{\link{natural_abundance_correction}}
}

\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/XiaoyangSu/AccuCor}
  \item Report bugs at \url{https://github.com/XiaoyangSu/AccuCor/issues}
}

}
\author{
\strong{Maintainer}: Lance Parsons \email{lparsons@princeton.edu} (\href{https://orcid.org/0000-0002-8521-714X}{ORCID})

Authors:
\itemize{
  \item Xiaoyang Su \email{xs137@rwjms.rutgers.edu} (\href{https://orcid.org/0000-0001-8081-1396}{ORCID})
}

Other contributors:
\itemize{
  \item Yujue Wang \email{yw429@rwjms.rutgers.edu} (\href{https://orcid.org/0000-0001-7088-1729}{ORCID}) [contributor]
  \item Princeton University [copyright holder]
}

}
