% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucor_lib.R
\name{carbon_correction}
\alias{carbon_correction}
\title{Natural Abundance correction for Carbon labeled samples}
\usage{
carbon_correction(InputFile, InputSheetName = "Sheet1",
  ColumnsToSkip = NULL, Resolution, ResDefAt, C13Purity = 0.99,
  ReportPoolSize = TRUE)
}
\arguments{
\item{InputFile}{Input xlsx file.}

\item{InputSheetName}{Name of sheet in xlsx file with columns 'compound',
'formula', 'isotopelabel', and one column per sample.}

\item{ColumnsToSkip}{Specify column heading to skip. All other columns not
named 'compound', 'formula', and 'isotopelabel' will be assumed to be
sample names.}

\item{Resolution}{For Exactive, the Resolution is 100000, defined at Mw 200}

\item{ResDefAt}{Resolution defined at (in Mw), e.g. 200 Mw}

\item{C13Purity}{Carbon 13 purity, default: 0.99}

\item{ReportPoolSize}{default: TRUE}
}
\value{
Named list of matrices: 'Corrected', 'Normalized',
     'PoolBeforeDF', and 'PoolAfterDF'.
}
\description{
Natural Abundance correction for Carbon labeled samples
}
\examples{
\dontrun{
carbon_correction("ExcelFile", Resolution=100000, ResDefAt=200)
}
}
