% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{clean_data_frame}
\alias{clean_data_frame}
\title{Standardize data frame columns and data types}
\usage{
clean_data_frame(df, columns_to_skip = NULL)
}
\arguments{
\item{df}{Data frame to clean}

\item{columns_to_skip}{Specify column heading to skip. All other columns not
named 'compound', 'formula', and 'isotopelabel' will be assumed to be
sample names.}
}
\value{
"cleaned" data.frame which with columns 'compound', 'formula',
     'isotope_label', label_index', followed by columns for each sample
}
\description{
Standardize data frame columns and data types
}
