% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucor_lib.R
\name{natural_abundance_correction}
\alias{natural_abundance_correction}
\title{Natural Abundance correction for specified isotope}
\usage{
natural_abundance_correction(path, sheet = NULL, output_path = NULL,
  ColumnsToSkip = NULL, Resolution, ResDefAt, purity = 0.99,
  ReportPoolSize = TRUE)
}
\arguments{
\item{path}{Path to xlsx file.}

\item{sheet}{Name of sheet in xlsx file with columns 'compound',
'formula', 'isotopelabel', and one column per sample.}

\item{output_path}{Path to output file, default is input path with
`_corrected` appended. If `FALSE` then no output file is written.}

\item{ColumnsToSkip}{Specify column heading to skip. All other columns not
named 'compound', 'formula', and 'isotopelabel' will be assumed to be
sample names.}

\item{Resolution}{For Exactive, the Resolution is 100000, defined at Mw 200}

\item{ResDefAt}{Resolution defined at (in Mw), e.g. 200 Mw}

\item{purity}{Isotope purity, default: 0.99}

\item{ReportPoolSize}{default: TRUE}
}
\value{
Named list of matrices: 'Corrected', 'Normalized',
     'PoolBeforeDF', and 'PoolAfterDF'.
}
\description{
Natural Abundance correction for specified isotope
}
\examples{
\dontrun{
natural_abundance_correction("ExcelFile", Resolution=100000, ResDefAt=200)
}
}
