% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucor.R
\docType{package}
\name{accucor}
\alias{accucor}
\title{accucor: A package for natural abundance correction of mass spectrometer data}
\description{
AccuCor is an isotope natural abundance correction algorithm that
is needed especially for high resolution mass spectrometers. AccuCor
supports correction for 13C, 2H and 15N.
}
\section{AccuCor functions}{

\code{\link{natural_abundance_correction}}
}

