% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/accucor_lib.R
\name{natural_abundance_correction}
\alias{natural_abundance_correction}
\title{Natural Abundance correction for mass spectrometry data}
\usage{
natural_abundance_correction(path, sheet = NULL, output_base = NULL,
  columns_to_skip = NULL, resolution, resolution_defined_at, purity = 0.99,
  report_pool_size = TRUE)
}
\arguments{
\item{path}{Path to xlsx file.}

\item{sheet}{Name of sheet in xlsx file with columns 'compound',
'formula', 'isotopelabel', and one column per sample.}

\item{output_base}{Path to basename of output file, default is input path.
`_corrected` will be appended to basename. Filetype is determined by
file extension.  If `FALSE` then no output file is written.}

\item{columns_to_skip}{Specify column heading to skip. All other columns not
named 'compound', 'formula', and 'isotopelabel' will be assumed to be
sample names.}

\item{resolution}{For Exactive, the Resolution is 100000, defined at Mw 200}

\item{resolution_defined_at}{Resolution defined at (in Mw), e.g. 200 Mw}

\item{purity}{Isotope purity, default: 0.99}

\item{report_pool_size}{default: TRUE}
}
\value{
Named list of matrices: 'Corrected', 'Normalized',
     'PoolBeforeDF', and 'PoolAfterDF'.
}
\description{
\code{natural_abundance_correction} returns the corrected and normalized
intensities of isotopically labeled mass spectrometry data. It was designed
to work with input data from
\href{https://elucidatainc.github.io/ElMaven/}{El-MAVEN} and
\href{http://maven.princeton.edu}{MAVEN} software.
}
\details{
C13, H2, and N15 isotopes are supported. The isotopes are detected from the
\code{isotopeLabel} column of the input file. The expected label text is
\code{C13-label-#}. \code{D-label-#}. or \code{N15-label-#}. Parent
(unlabeled) compounds are specified by \code{C12 PARENT}.
}
\examples{
\dontrun{
natural_abundance_correction("inst/extdata/C_Sample_Input_Simple.xlsx",
                             Resolution=100000, ResDefAt=200)
}
}
