% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/io_functions.R
\name{read_elmaven_xlsx}
\alias{read_elmaven_xlsx}
\title{Natural Abundance correction for Carbon labeled samples}
\usage{
read_elmaven_xlsx(path, sheet = NULL, ColumnsToSkip = NULL, ...)
}
\arguments{
\item{path}{Path to xlsx file.}

\item{sheet}{Name of sheet in xlsx file with columns 'compound',
'formula', 'isotopelabel', and one column per sample.}

\item{ColumnsToSkip}{Specify column heading to skip. All other columns not
named 'compound', 'formula', and 'isotopelabel' will be assumed to be
sample names.}

\item{\dots}{Pass additional parameters to readxl::read_excel}
}
\value{
List with two dataframes, "original" which is result of read_excel
      and "cleaned", which is a dataframe with columns 'compound', 'formula',
      'isotope_label', label_index', followed by columns for each sample.
}
\description{
Natural Abundance correction for Carbon labeled samples
}
\examples{
\dontrun{
read_elmaven_xlsx("ExcelFile", "Sheet1")
}
}
